add amendment caching
